pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = "Habit Tracker"
authors = [
    {name = "Станислав Маяцкий",email = "stasm226@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pre-commit (>=4.2.0,<5.0.0)",
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "ipython (>=9.1.0,<10.0.0)"
]

[tool.poetry]
packages = [{include = "habit_tracker", from = "src"}]


[tool.poetry.group.dev.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
django-stubs = "^5.1.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
known_django = ["django"]
known_first_party = ["apps", "habit_tracker"]
known_third_party = ["rest_framework"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip_gitignore = true
skip = ["migrations"]

[tool.flake8]
max-line-length = 119
extend-ignore = ["E203", "W503"]
exclude = ".git,__pycache__,migrations,venv,.venv"
max-complexity = 10

[tool.mypy]
exclude = 'venv'
# Игнорировать отсутствующие типы у сторонних библиотек
ignore_missing_imports = true
# Поддержка strict режима, но без фанатизма
disallow_untyped_defs = false
check_untyped_defs = true
show_error_codes = true
follow_imports = "silent"
# Указать путь к settings, чтобы plugin понимал проект
django_settings_module = "config.settings"
